requirements.txt
PyPDF2==3.0.1
python-docx==1.1.0
pypandoc==1.12
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Werkzeug==3.0.1
